feat: add ParticleField component for dynamic particle effects; integrate into ThreeDCubes and ThreeDPlayground for enhanced visual experience with customizable particle properties.
